requirements.txt
alembic==1.4.2
aniso8601==8.0.0
attrs==19.3.0
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
coverage==5.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-CLI==0.4.0
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
gunicorn==20.0.4
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jose==1.0.0
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.6.1
more-itertools==8.4.0
numpy==1.19.0
packaging==20.4
pandas==1.0.5
pep8==1.7.1
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.9.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.18
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.1.0
